# docs/api/openapi.json
{"openapi":"3.0.3","info":{"title":"API","version":"1.0.0","contact":{}},"paths":{"/agent/pool/stats":{"get":{"operationId":"get~agent_scheduler.pool_stats","summary":"\u83b7\u53d6\u5b9e\u4f8b\u6c60\u7edf\u8ba1\u4fe1\u606f","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/agent/pool/register":{"post":{"operationId":"post~agent_scheduler.register_agent","summary":"\u6ce8\u518c\u65b0 Agent \u5b9e\u4f8b","description":"Body: {\"name\": \"agent-shanghai\", \"base_url\": \"http://192.168.1.101:4096\"}","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/agent/tools":{"get":{"operationId":"get~agent_scheduler.get_tools","summary":"\u83b7\u53d6\u6240\u6709\u5de5\u5177","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}},"post":{"operationId":"post~agent_scheduler.register_tool","summary":"\u6ce8\u518c\u65b0\u5de5\u5177","description":"Body: {\"name\": \"tool_name\", \"description\": \"...\", \"input_schema\": {...}, ...}","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/agent/session":{"post":{"operationId":"post~agent_scheduler.create_session","summary":"\u521b\u5efa\u65b0\u4f1a\u8bdd","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/agent/skills":{"get":{"operationId":"get~agent_scheduler.get_skills","summary":"\u83b7\u53d6\u6240\u6709\u6280\u80fd","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}},"post":{"operationId":"post~agent_scheduler.register_skill","summary":"\u6ce8\u518c\u65b0\u6280\u80fd","description":"Body: {\"name\": \"skill_name\", \"description\": \"...\", \"triggers\": [...], ...}","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/agent/chat":{"post":{"operationId":"post~agent_scheduler.chat","summary":"\u53d1\u9001\u6d88\u606f\u5e76\u83b7\u53d6\u56de\u590d","description":"\u5982\u679c\u4e0d\u63d0\u4f9b session_id\uff0c\u5219\u521b\u5efa\u65b0\u4f1a\u8bdd\u3002","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/agent/chat/stream":{"post":{"operationId":"post~agent_scheduler.chat_stream","summary":"\u53d1\u9001\u6d88\u606f\u5e76\u83b7\u53d6\u6d41\u5f0f\u56de\u590d (SSE)","description":"\u5982\u679c\u4e0d\u63d0\u4f9b session_id\uff0c\u5219\u521b\u5efa\u65b0\u4f1a\u8bdd\u3002","tags":["agent_scheduler"],"responses":{"default":{"description":"OK"}}}},"/api-docs":{"get":{"operationId":"get~docs","summary":"Docs","responses":{"default":{"description":"OK"}}}},"/health":{"get":{"operationId":"get~health","summary":"\u5065\u5eb7\u68c0\u67e5\u7aef\u70b9","responses":{"default":{"description":"OK"}}}},"/ready":{"get":{"operationId":"get~ready","summary":"\u5c31\u7eea\u68c0\u67e5\u7aef\u70b9","responses":{"default":{"description":"OK"}}}},"/agent/session/{session_id}/messages":{"get":{"operationId":"get~agent_scheduler.get_messages","summary":"\u83b7\u53d6\u4f1a\u8bdd\u5386\u53f2\u6d88\u606f","tags":["agent_scheduler"],"parameters":[{"name":"session_id","schema":{"type":"string"},"required":true,"in":"path"}],"responses":{"default":{"description":"OK"}}}},"/agent/session/{session_id}":{"delete":{"operationId":"delete~agent_scheduler.delete_session","summary":"\u5220\u9664\u4f1a\u8bdd","tags":["agent_scheduler"],"parameters":[{"name":"session_id","schema":{"type":"string"},"required":true,"in":"path"}],"responses":{"default":{"description":"OK"}}},"get":{"operationId":"get~agent_scheduler.get_session","summary":"\u83b7\u53d6\u4f1a\u8bdd\u4fe1\u606f","tags":["agent_scheduler"],"parameters":[{"name":"session_id","schema":{"type":"string"},"required":true,"in":"path"}],"responses":{"default":{"description":"OK"}}}},"/agent/tools/{name}/enabled":{"patch":{"operationId":"patch~agent_scheduler.update_tool_enabled","summary":"\u542f\u7528\u6216\u7981\u7528\u5de5\u5177","description":"Body: {\"enabled\": true/false}","tags":["agent_scheduler"],"parameters":[{"name":"name","schema":{"type":"string"},"required":true,"in":"path"}],"responses":{"default":{"description":"OK"}}}},"/agent/pool/{name}":{"delete":{"operationId":"delete~agent_scheduler.unregister_agent","summary":"\u6ce8\u9500 Agent \u5b9e\u4f8b","tags":["agent_scheduler"],"parameters":[{"name":"name","schema":{"type":"string"},"required":true,"in":"path"}],"responses":{"default":{"description":"OK"}}}},"/agent/session/{session_id}/abort":{"post":{"operationId":"post~agent_scheduler.abort_session","summary":"\u4e2d\u6b62\u8fd0\u884c\u4e2d\u7684\u4f1a\u8bdd","tags":["agent_scheduler"],"parameters":[{"name":"session_id","schema":{"type":"string"},"required":true,"in":"path"}],"responses":{"default":{"description":"OK"}}}}},"tags":[{"name":"agent_scheduler"}],"servers":[],"security":[]}